# docs/swagger.yaml
basePath: /
...
info:
  contact:
    email: vvbbnn00@gmail.com
    name: vvbbnn00
    url: https://github.com/vvbbnn00/goflet
  description: Goflet is a lightweight file upload and download service written in
    Go.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Goflet API
  version: unknown
paths:
...
  /file/{path}:
    delete:
      description: Delete a file by path, {path} should be the relative path of the
        file, starting from the root directory, e.g. /file/path/to/file.txt
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      responses:
        "204":
          description: Deleted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: File not found or upload not started
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: Delete File
      tags:
      - File
    get:
      description: Download a file by path, supports range requests, {path} should
        be the relative path of the file, starting from the root directory, e.g. /file/path/to/file.txt
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          headers:
            Cache-Control:
              description: max-age=3600
              type: string
            Content-Disposition:
              description: attachment; filename=file.txt
              type: string
            Content-Length:
              description: "1024"
              type: string
            Content-Type:
              description: application/octet-stream
              type: string
            ETag:
              description: 686897696a7c876b7e
              type: string
            Last-Modified:
              description: Mon, 02 Jan 2006 15:04:05 GMT
              type: string
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: File Download
      tags:
      - File
    head:
      description: Download a file by path, supports range requests, {path} should
        be the relative path of the file, starting from the root directory, e.g. /file/path/to/file.txt
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          headers:
            Cache-Control:
              description: max-age=3600
              type: string
            Content-Disposition:
              description: attachment; filename=file.txt
              type: string
            Content-Length:
              description: "1024"
              type: string
            Content-Type:
              description: application/octet-stream
              type: string
            ETag:
              description: 686897696a7c876b7e
              type: string
            Last-Modified:
              description: Mon, 02 Jan 2006 15:04:05 GMT
              type: string
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: File Download
      tags:
      - File
    post:
      consumes:
      - multipart/form-data
      description: Upload a small file using a POST request, {path} should be the
        relative path of the file, starting from the root directory, e.g. /file/path/to/file.txt
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      - description: File
        in: formData
        name: file
        required: true
        type: file
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: File not found or upload not started
          schema:
            type: string
        "409":
          description: File completion in progress
          schema:
            type: string
        "413":
          description: File too large, please use PUT method to upload large files
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: Upload Small File
      tags:
      - File
      - Upload
  /upload/{path}:
    delete:
      description: Cancel an upload session, {path} should be the relative path of
        the file, starting from the root directory, e.g. /upload/path/to/file.txt
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      responses:
        "204":
          description: Deleted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Upload session not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Cancel Upload
      tags:
      - Upload
    post:
      description: Complete an upload session with a partial file upload. You should
        first upload the file with a PUT request, then complete the upload with a
        POST request, {path} should be the relative path of the file, starting from
        the root directory, e.g. /file/path/to/file.txt
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: File not found or upload not started
          schema:
            type: string
        "409":
          description: File completion in progress
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: Complete Partial File Upload
      tags:
      - Upload
    put:
      consumes:
      - '*/*'
      description: Create an upload session with a partial file upload, supports range
        requests, {path} should be the relative path of the file, starting from the
        root directory, e.g. /file/path/to/file.txt
      parameters:
      - description: File path
        in: path
        name: path
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "403":
          description: Directory creation not allowed
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Authorization: []
      summary: Partial File Upload
      tags:
      - Upload
securityDefinitions:
  Authorization:
    description: 'You need to provide a valid jwt token in the header, in headers,
      you should provide a key-value pair like this: Authorization: Bearer xxxxxx;
      The token has the same effect as the token in the query string, but it is more
      secure than the token in the query string. Or you can just provide the token
      in the query string, like this: ?token=xxxxxx. More info about jwt: https://github.com/vvbbnn00/goflet?tab=readme-ov-file#authentication-method'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"